Create Module for Team search methods Include Team Search module
