+++ lib/stat_tracker.rb
@@ -3,9 +3,11 @@ require_relative './team'
 require_relative './season'
 require_relative './game'
 require_relative './merge_sortable'
+require_relative './team_searchable'
 
 class StatTracker
   include MergeSort
+  include TeamSearchable
 	attr_reader :seasons, :teams, :game_teams
 
 	def initialize(locations)
@@ -28,27 +30,6 @@ class StatTracker
 
 	def count_of_teams
 		@teams.length
-	end
-
-  def highest_scoring_visitor
-    # merge_sort(@teams, "average_goals_away")[-1].team_name
-		@teams.max_by(&:average_goals_away).team_name
-	end
-
-  #alex
-  def highest_scoring_home_team
-    # merge_sort(@teams, "average_goals_home")[-1].team_name
-		@teams.max_by(&:average_goals_home).team_name
-	end
-
-  def lowest_scoring_visitor
-    # merge_sort(@teams, "average_goals_away")[0].team_name
-    @teams.min_by(&:average_goals_away).team_name
-	end
-
-  def lowest_scoring_home_team
-    # merge_sort(@teams, "average_goals_home")[0].team_name
-		@teams.min_by(&:average_goals_home).team_name
 	end
 
 	def count_of_games_by_season
@@ -97,34 +78,5 @@ class StatTracker
 			acc[season.id.to_s] = (season.games_unsorted.sum {|game| game.total_score}/season.games_unsorted.length.to_f).round(2)
 			acc
 		end
-	end
-
-	def best_offense
-		@teams.max_by { |team| team.average_goals_total }.team_name
-	end
-
-	def worst_offense
-		@teams.min_by { |team| team.average_goals_total }.team_name
-	end
-
-	def winningest_team
-		@teams.max_by { |team| team.win_percent_total }.team_name
-	end
-
-	def worst_fans
-		worst_fan_teams = @teams.find_all { |team| team.average_goals_away > team.average_goals_home}
-		worst_fan_teams.map(&:team_name)
-	end
-
-	def best_fans
-		@teams.max_by { |team| (team.home_win_percentage - team.away_win_percentage)}.team_name
-	end
-
-	def best_defense
-		@teams.min_by { |team| team.total_scores_against }.team_name
-	end
-
-	def worst_defense
-		@teams.max_by { |team| team.total_scores_against }.team_name
 	end
 end
